+++ src/main/java/br/com/fiap/controlepedidos/adapters/driver/apirest/dto/in/MercadoPagoQRCodePaymentCallbackDTO.java
@@ -1,11 +1,15 @@
 package br.com.fiap.controlepedidos.adapters.driver.apirest.dto.in;
 
+import br.com.fiap.controlepedidos.adapters.driver.apirest.dto.ProductDTO;
+import br.com.fiap.controlepedidos.core.domain.entities.Product;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record MercadoPagoQRCodePaymentCallbackDTO(@NotNull(message = ORDER_ID_REQUIRED) UUID orderId,
-                                                  @NotNull(message = PAID_VALUE_REQUIRED) float paidValue) {
+                                                  @NotNull(message = PAID_VALUE_REQUIRED) BigDecimal paidValue) {
     private static final String ORDER_ID_REQUIRED = "ID da ordem não pode ser nulo";
     private static final String PAID_VALUE_REQUIRED = "O valor pago não foi informado";
 }
+
