+++ src/main/java/br/com/fiap/controlepedidos/core/application/services/checkout/ConfirmQRCodePaymentOrder.java
@@ -5,5 +5,5 @@ import br.com.fiap.controlepedidos.core.domain.entities.Payment;
 import java.util.UUID;
 
 public interface ConfirmQRCodePaymentOrder {
-    public Payment confirmQrCodePayment(UUID orderId, float paidValue) throws Exception;
+    public Payment confirmQrCodePayment(UUID orderId, int paidValue) throws Exception;
 }
