+++ src/main/java/br/com/fiap/controlepedidos/core/application/services/order/PayOrderService.java
@@ -7,6 +7,6 @@ import java.util.UUID;
 
 public interface PayOrderService {
 
-    Payment payOrder(Order order, float receivedValue) throws Exception;
+    Payment payOrder(Order order, int receivedValue) throws Exception;
 
 }
