+++ src/main/java/br/com/fiap/controlepedidos/core/application/services/order/impl/PayOrderServiceImpl.java
@@ -21,7 +21,7 @@ public class PayOrderServiceImpl implements PayOrderService {
     }
 
     @Override
-    public Payment payOrder(Order order, float receivedValue) throws Exception {
+    public Payment payOrder(Order order, int receivedValue) throws Exception {
 
         boolean result = order.payOrder(receivedValue);
